refactor(crypto): CRP-700 cleanup ic-request(-auth-delegation) domain separators (#7635)
Performs a cleanup regarding the `ic-request` and
`ic-request-auth-delegation` domain separators:
* Moves `DOMAIN_IC_REQUEST`, which is used for _signing_ ingress
messages, from `ic_types::crypto::hash` to the better suited
`ic_types::crypto::sign` module. Dependent code is not impacted because
there is a re-export in `ic_types::crypto`, which is also adapted.
* Removes the obsolete `SIG_DOMAIN_IC_REQUEST` constant and instead uses
the already existing `DomainSeparator::IcRequest`.
* Removes the `SIG_DOMAIN_IC_REQUEST_AUTH_DELEGATION` constant and
instead uses the newly created
`DomainSeparator::IcRequestAuthDelegation`, which also has the advantage
that uniqueness of that variant is checked automatically.