feat: Testing Canister for testing Message Routing in State Machine Tests (#7551)
[This is a resubmit of #7325, which got automatically closed.]

Previous canisters used for this purpose are either for a very specific
purpose (`downstream-calls-test-canister`) or based on generating random
traffic, i.e. `xnet-test-canister` (designed for system tests, not state
machine tests) and `random-traffic-test-canister`. The latter is an
attempt at better following what is happening inside the state machine
by gathering internal logs.
The main problem comes from the fact that a canister based on random
traffic cannot be instructed to do something specific, rather one must
generate traffic until a condition is fulfilled, and from the fact that
the traffic is generated from the heartbeat method. Since the heartbeat
method does not return the outcome of a call to the controller of the
state machine, it can be difficult to understand what is going on.
Additionally, turning the traffic on and off is done through an ingress
message, but executing this triggers more heartbeats. This implicit
execution can be quite confusing.

The canister in this PR works in a different way:
- It makes calls exactly as instructed, and then downstream calls from
there and so on.
- It does so triggered from an ingress message, which will return the
outcome of the call. There is no implicit execution and exactly what is
enqueued in the ingress pool gets executed and nothing else.
- Random traffic is produced by generating random instructions (and then
random downstream calls from there and so forth) which are then sent to
a canister by an ingress message, enqueued in the pool and eventually
executed.

Overall this canister can still produce randomized traffic through
randomly generated instructions, but it allows better control over what
is executed and it provides direct feedback for each call it made (or
failed to perform).

One pitfall to keep in mind though: Ingress messages in general do not
execute in the order they are enqueued, but this can be enforced by
enqueuing, then executing rounds until it is in `Processing` state and
only then enqueuing the next ingress message.

---------

Co-authored-by: Christian Stieger <christian.stieger@dfinity.org>