refactor(crypto): CRP-543 make Send + Sync explicit for some Crypto traits (#7634)
The traits relating to the Crypto components are all designed to support
multi-threading, but this is not apparent in the trait definition. This
cleanup PR make the multi-thread support explicit for some selected
crypto traits by adding the `Send` and `Sync` trait bounds, so that it
is not necessary to specify these trait bounds over and over again
during usage.

In particular, this is done for
* `ic_interfaces::crypto::Crypto`
* `ic_interfaces::crypto::BasicSigner`
* `ic_crypto_interfaces_sig_verification::IngressSigVerifier`
* `ic_crypto_tls_interfaces::TlsConfig`

because these traits are explicitly used outside of the Crypto code
(i.e., outside of `rs/crypto`) with the `Send` and `Sync` trait bounds.