refactor(nns): Fix places where ManageNeuron is incorrectly used (#7618)
# Why

The `ManageNeuron` API types have 2 sets of types - ones that used for
requests and one that uses for responses. In some places we are misusing
the ones for the responses to send requests.

Note: This is error-prone, but there is a path towards clarity:
- Have the [next PR](https://github.com/dfinity/ic/pull/7619) clarify
that the purpose of the type is a "manage neuron proposal"
- Once we no longer exposes proposal actions as responses we can clean
up the `ManageNeuronProposal` (renamed in this PR) and related types