refactor(nns): Rename ManageNeuron to ManageNeuronProposal (#7619)
# Why

The `ManageNeuron` concept is used in 2 different contexts: (1) a
`manage_neuron` canister method, and (2) a "neuron management proposal".
There are also 2 kinds of API types: one that represents requests and
one that represent responses (they are complicated because of historic
reasons, but we will have a chance to significantly simplify them by
using a self-describing type for responses). For the API types that
represents responses (`pb::ManageNeuron` and
`pb::manage_neuron::Command`), they are only ever useful for the neuron
management proposals, as we shouldn't need to represent a ManageNeuron
command in API responses otherwise. Therefore, it's better to
disambiguate the type to reflect this reality. This PR also aims at
prevent the mistakes fixed in a [previous
PR](https://github.com/dfinity/ic/pull/7618)

# What

- Rename `pb::ManageNeuron` -> `pb::ManageNeuronProposal`
- Rename `pb::manage_neuron::Command` -> `pb::manage_neuron::
ManageNeuronProposalCommand`