perf(crypto): CRP-2341 CRP-1621 Optimize the NIDKG sharing proof (#7671)
Some small optimizations for the NIDKG sharing proofs:
* In the second verification equation, we can save a group
multiplication in the computation of LHS if we initialize ik to
x_challenge instead of 1.
* In the third verification equation: 
* We cans ave the additional scalar multiplication by z_r (as part of
mul2) by computing scalars z_r*x^i and using them in the multi-scalar
multiplication on the LHS..
* We can rearrange the verification equation to perform a single
multi-scalar multiplication instead of one on the LHS and one on the
RHS.




On my laptop, this improves sharing proof verification by about 20%
(goes from 10.3 ms to 8.6 ms).

---------

Co-authored-by: Andrea Cerulli <19587477+andreacerulli@users.noreply.github.com>