feat(node-rewards-canister): Add versioning on rewards calculation algorithm (#7705)
This pull request introduces versioning support for the node rewards
calculation algorithm. The main goal is to allow clients to specify
which version of the rewards calculation algorithm to use, and to make
the API forward-compatible as new algorithm versions are introduced. The
changes add a `RewardsCalculationVersion` struct, update API request and
response types to include the version, and refactor the calculation
logic to support multiple algorithm versions (currently only v1 is
implemented).