test(crypto): CRP-2950 extend request signature system tests for all api versions and read_state (#7613)
Extends the `rs/tests/crypto/request_signature_test.rs` system test to
ensure that for all failure scenarios we test all api versions and also
read state requests (in addition to queries and updates).

In the future, we should consider creating better test infrastructure so
that the tests can be written in a more concise way.