test(Rosetta): DEFI-2452: Remove local cluster duplicate monitoring (#7679)
The script that deploys a local k8s cluster for Rosetta testing installs
two monitoring stacks simultaneously:

- A prometheus Helm release (older/existing)
- A kube-prometheus Helm release (newer/conflicting)

This causes a race condition, where sometimes the Grafana instance is
not available on port 3000.

This PR proposes to remove the older prometheus and Grafana setups, and
adjust the configuration of the newer ones to match the old ones, so
that the dashboards and instructions stay the same.