test(crypto): CRP-2950 extend system tests to ensure invalid signatures fail (#7611)
Extends the `rs/tests/crypto/request_signature_test.rs` system test to
ensure that updates, queries, and read state requests fail for all the
different API versions if the signature is invalid (repeated bytes of
42).

As sanity check, I tested not invalidating the signature and then all
requests return with status 2xx.
In the future, we should consider creating better test infrastructure so
that the tests can be written in a more concise way.