chore: run ingress filter on latest certified state (#7685)
This PR makes the ingress filter run on the latest certified state
(instead of the latest - potentially uncertified - state).

Motivation
---

This change fixes the following inconsistency for non-replicated
execution:
- A non-replicated query call uses the latest certified state (necessary
because a certificate is provided to execution).
- A canister http outcall transform uses the latest certified state.
- A canister ingress filter uses the latest (potentially uncertified)
state!

In the last two cases, no certificate is provided to execution and thus
both certified and uncertified states could be used technically:
- the latest certified state is more "trustworthy" and less prone to
divergence;
- the latest (potentially uncertified) state is more "recent".

Since it is not clear why a more "recent" uncertified state is more
appropriate, we switch to using the latest certified state in all cases.

Changes to Tests
---

StateMachine/PocketIC tests execute an empty round at the very beginning
to always have a certified state available. This makes the first
checkpoint be produced earlier and changes the randomness delivered to
canisters in such tests (its seed is derived from the round number which
increases by one). There's no change in round timestamps.