test: Promote the refunds_test queue compatibility test to bidirectional
When the new certification version allowing for encoding refunds in streams was introduced, the included refunds_test had to run as a SelfTestOnly, until the certification version had made it onto all of mainnet. Now that this has been the case for a while, promote it to a bidirectional test, that tests roundtrip encoding between the HEAD code and the mainnet code.