fix: Fix running state machine tests with `cargo` (#7701)
As things stand, most state machine tests cannot be run from `cargo`
(and, by extension, cannot be debugged from an IDE). There are a couple
of root causes for this:
1. Building the various sandbox binaries fails because of outdated
paths. This is easily fixed.
2. Canister binaries are built without a `--package` argument, so their
dependencies get a superset of all the features needed by any crate in
the workspace. This causes the resulting Wasm module to fail to install
because e.g. `ic_cdk` ends up pulling in `js_sys` and `getrandom`. The
fix is to specify a fully qualified canister binary name, i.e. a
combination of crate/package name and binary name, so the minimal set of
features is enabled (just as Bazel does).

I have only done the latter for one test and one canister, as getting it
to work is tedious work: one needs to find the crate name; and more
often than not, add an `ic-dummy-getrandom-for-wasm` dependency (which
Bazel does implicitly for all canister binaries).