feat(outcalls): make app subnets use API boundary nodes (#7586)
The crux of the change is in
[pool_manager.rs](https://github.com/dfinity/ic/pull/7586/files#diff-443cc7cc0722e2b3cdcb757bd5aacd228f516450ac5af15d35152b6d383993f3),
where instead of sending an empty list to the http adapter, we instead
send the list of API boundary nodes that are assigned to proxy nodes on
app subnets.

The HTTP adapter will then proxy requests via API bns in the request if
the direct request fails.

The socks proxy url config parameter was also removed with the change.