fix(IDX): set ref for paths-filter (#7730)
According to the [docs](https://github.com/dorny/paths-filter) the ref
always defaults to `${{ github.ref }}` - but since we are now calling it
from workflow_dispatch for external contributors we need to set this
explicitly.

I believe we don't need to set `base` since it's set to the default
branch of the repo but would appreciate a second set of eyes to
determine if we do in fact need to set it to `base_ref` like we do for
didc checks.

Here is what's in the docs for reference:
```
    # Branch, tag, or commit SHA against which the changes will be detected.
    # If it references the same branch it was pushed to,
    # changes are detected against the most recent commit before the push.
    # Otherwise, it uses git merge-base to find the best common ancestor between
    # current branch (HEAD) and base.
    # When merge-base is found, it's used for change detection - only changes
    # introduced by the current branch are considered.
    # All files are considered as added if there is no common ancestor with
    # base branch or no previous commit.
    # This option is ignored if action is triggered by pull_request event.
    # Default: repository default branch (e.g. master)
    base: ''
```