chore: switch NTP to NTS (#7686)
# NODE-1782

Switched the chrony config to "authenticated" NTP. Picked the NTS
sources from https://github.com/jauderho/nts-servers/ and tested them
manually:
```
chronyd -Q -t 3 'server NTP_SERVER_URL nts iburst maxsamples 1'
```
As a fallback, I kept the unauthenticated pools around as fallback. We
will remove the fallback pool once we have enough confidence in the NTS
servers.

In addition, we had to allow traffic go to port 4460 on remote addresses
as this is what is used for the NTS key-exchange. This is required in
addition to port 123 for NTP.