refactor: adding options to allow all without specifying all of the members in whitelisting for node swapping feature flags (#7650)
The PR refactors how empty lists are interpreted for the node swapping
feature.

Previously an explicit statement of *who* is allowed to swap the nodes
and *where* was needed. While this approach is good and it works fine,
there is an issue that we have a lot of subnets and a lot of node
operators:
* we have 47 subnets
* we have 143 node operators

Having to state all of the subnets and all of the node operators will
make the `flags.rs` really unreadable. Furthermore, having all of them
stated doesn't bring any value.

*NOTE*: it will still be possible to disallow the feature on specific
subnets or for specific node operators. We will just have to make
explicit lists *without* the said subnets or node operators.

In the future, we may look into turning whitelisting model into a
blacklisting one if we see that we have to disallow feature for specific
node operators or for specific subnets.