refactor(nns): Define ValidProposalAction with all protobuf types (#7674)
The proposal validation logic is spread out in different places, and
every time the proposal action is accessed, the caller either needs to
make an assumption on the validity of the proposal (which can be far
away from the callsite and hence can easily be broken), or validate it
again. The `Valid*` pattern makes it so that the validation can be
consolidated into the type, and also avoid mistakes such as forget to
validate the type in some places. In the future, each action can be
changed into Valid*Action types, and this PR makes the minimum change to
allow the future refactoring.