fix(ckdoge): use correct transaction fees  (#7660)
Follow-up on #7360 to use the correct parameters to compute the fees
involved in a transaction sent by the minter to the Dogecoin network in
case of a withdrawal.

The implementation introduces a new trait `FeeEstimator`, of which there
are two implementations (one for Bitcoin and another one for Dogecoin),
and parametrize the existing methods to build transactions by an
instance of that trait. A follow-up PR will add the
`estimate_withdrawal_fee` endpoint.