feat(nns): Define and use ValidExecuteNnsFunction (#7724)
Every time an ExecuteNnsFunction is used, it most likely needs to be
validated (e.g. nns function id is valid and is not obsolete). A
`ValidExecuteNnsFunction` type consolidates the validation logic and
make it impossible to represent an invalid one with this type. This will
also help with transforming into a self-describing proposal action.