test: drop requires-network from PocketIC tests (#7753)
This PR removes the `requires-network` tag from PocketIC tests.

This tag was introduced long ago to mitigate spurious failures to bind
to 127.0.0.1 when creating an HTTP gateway on Apple Sillicon. This PR
removes the tag to reevaluate if it still needed. If we still observe
failures, this PR will be reverted.