fix: flaky DTS test (#7771)
It can sometimes happen that the background tasks of the state manager
outlive a test, resulting in the state manager getting the on-disk data
wiped under its feet, which, in turn, can lead to flakiness.

This PR attempts to fix the flakiness we observe in the DTS test by
adding an additional step to `set_checkpoints_enabled` to wait for all
asynchronous operations to finish before returning control to the
caller.