feat(ICRC-Index): add and handle ICRC-107 fee collector blocks (#7411)
Introduce the ICRC-107 fee collector operation to the ICRC ledger blocks
and implement handling of the operation in the ICRC index. Since the
ledger does not produce the new blocks yet, the index handling is tested
using the icrc3 test ledger. The blocks structure and handling is
implemented according the the ICRC-107 standard as defined in
https://github.com/dfinity/ICRC/pull/117. Although the standard allows
for 107 blocks to skip the `tx` entirely, we require the `tx` and the
field `tx.op` to be present in the block.

---------

Co-authored-by: Mathias Björkqvist <mathias.bjorkqvist@dfinity.org>