chore: use RoutingTable struct to create new subnets dynamically (#7779)
This PR makes PocketIC use a (hard-coded) routing table of the ICP
mainnet to derive the canister ranges of an ICP mainnet subnet to which
a canister ID belongs instead of deriving the canister ranges by masking
bits in the binary representation of the canister ID.

This also allows us to create the II subnet (which has been split from
the NNS subnet and thus its canister ranges cannot be derived by masking
bits) dynamically after a PocketIC instance is created and an existing
test is extended to cover this new case.

The motivation for this change is that canister migration will make it
impossible to derive canister ranges of ICP mainnet subnets by masking
bits in the binary representation of canister IDs.

For now, the ICP mainnet routing table was manually obtained by running
`ic-admin get-routing-table --json` (after a fix from this
[PR](https://github.com/dfinity/ic/pull/7781)) and commiting the JSON
representation of the ICP mainnet routing table as the file
`/rs/pocket_ic_server/src/mainnet_routing_table.json`. This flow might
improve in the future.