refactor: [CON-1551] [CON-1552] move registry client polling logic inside the replicator (#5909)
When the registry replicator is created, it creates a registry client
that instantly starts polling the local store to update its cache
([source](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/orchestrator/registry_replicator/src/lib.rs?L108-110)).
It is only when calling `start_polling()` that the underlying data (the
local store) is initialized
([source](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/orchestrator/registry_replicator/src/lib.rs?L316-319)).
This means the registry client will return empty data for the first 5
seconds.
Also, the registry replicator and client are independently updating the
local store and cache respectively every 5 seconds, which means a change
in the registry canister is effectively mirrored on nodes after a
potential delay of 10 seconds.

This PR proposes to refactor the external API of the RegistryReplicator
such that it abstracts away the underlying local store and registry
client, making it impossible to build a replicator with existing such
instances. That way, the replicator can initialize the local store in
its constructor, and initialize the registry client without making it
start polling. In fact, the registry client will never poll on its own:
it will manually be polled just after the replicator does so. A
`RegistryReplicator` instance will thus always be initialized with an
initialized local store and its `RegistryClient`'s cache will always
reflect the latest data on disk (modulo rare race conditions, where we
read from the cache between the poll of the replicator and the poll of
the client). This means a change in the registry canister is mirrored on
nodes after maximum 5 seconds instead of 10.

This also has the advantage that external code using the
`RegistryReplicator` do not need to handle the polling of the registry
client themselves, thus reducing complexity (see how we can remove
local_store and/or registry_client dependencies in the
backup/ic_boundary/recovery crates)


https://github.com/dfinity/ic/pull/5909/commits/35b3cb1fac41462f8756e6198060a808dfa03b41
also modifies the internal state by
- Passing the config public key in the constructor, which allows to use
the fallback registry (i.e. from the config) in case the NNS URL/key
cannot be found in the registry, instead of returning Ok() early.
- Returning an error if the NNS URL/key cannot be found neither in the
registry nor in the config instead of returning Ok() early.

Finally, this PR also adds unit tests to the registry replicator.