chore: support parameterizing the NODES_COUNT in the tecdsa performance tests (#7815)
This moves the definition of the default number of nodes to use in the
application subnet of the tecdsa performance tests from Rust to Bazel.
This allows using a different number of nodes per test. Additionally it
allows overriding this number by passing `--test_env NODES_COUNT=40` to
`bazel test`.