chore: Fix flaky VetKD unknown contexts test (#7822)
In this test, we are attempting to validate a VetKD payload with
agreements for requests 3, 4 and 5. However, the replicated state only
contains contexts for requests 0, 1 and 2. Therefore, the payload should
be rejected.

Before https://github.com/dfinity/ic/pull/7373 introduced
multi-threading to the VetKD payload builder, the reason that the
payload was rejected was always because no context for the first
agreement (with ID 3) could be found. This is because each agreement was
checked sequentially. The test asserted that this was indeed the case.

After introducing multi-threading, any of the three agreements may be
checked first, which affects the rejection error. This PR fixes the
resulting flakiness by allowing any of the three requests to be part of
the rejection error.