feat(node-reward-canister): Added simulated call to mgm canister for tests (#7821)
This pull request adds support for a test-specific build of the node
rewards canister and implements a test mode for the
`node_metrics_history` method on the management canister whch simulate
the call. The main goal is to enable easier testing by providing
deterministic, mock data for node metrics queries when the `test`
feature is enabled.
This is mainly done because our `state_machine` tests do not support
multiple subnets