chore: ic-cdk v0.19 & ic-cdk-timers v1.0.0 (#7494)
# Description

Both `ic-cdk` and `ic-cdk-timers` rely on `ic-cdk-executor` for
asynchronous execution. Due to the strict requirement for a single
version of `ic-cdk-executor` within the dependency tree, the release of
`ic-cdk-executor` v2.0.0 necessitates a bulk dependency upgrade across
the monorepo.

`ic-cdk-executor` is considered stable now, i.e., future ic-cdk releases
will depend on `ic-cdk-executor `v2.0.0. This is projected to be the
last required unification of all `ic-cdk `dependencies to a single
version within the monorepo.

This PR also updates dependencies that depend on `ic-cdk`:
- `canbench` & `canbench-rs` v0.4.0
- `evm_rpc_client` v0.3.0
- `ic-canister-runtime` v0.1.0
- `ic-vetkeys` v0.6.0

Also updates:
- `ic-management-canister-types` v0.5.0, this is a dependency of
`ic-cdk` itself.
- `ic-canister-sig-creation` v1.3.1, the new version no longer depends
on `ic-cdk`.