feat: CON-1618 Add endpoint for requesting `SignedIDkgDealing` in block (#7698)
## Background
Currently, IDKG transcripts are created by the blockmaker, by
aggregating _at least_ `f+1` IDKG dealings with _at least_ `2f+1`
support shares each. These transcripts are included into the block,
transmitted to peers and verified as part of block validation.

## Problem
These IDKG transcripts can be quite large (up to 160KB). Consequently,
including many of them into the block increases the block size
significantly. This has an adverse effect on the finalization rate, as
it takes longer for large blocks to be transmitted. Furthermore, large
blocks consume bandwidth which otherwise could be used elsewhere, i.e.
for the IDKG protocol itself.

## Solution
The largest amount of data in each IDKG transcript (98.8%) is consumed
by the “[verified
dealings](https://sourcegraph.com/github.com/dfinity/ic@37a30ba73623f2b654515c0e942b30a229ae1c06/-/blob/rs/types/types/src/crypto/canister_threshold_sig/idkg.rs?L731)”,
which is a list of at least `f+1` batch-signed `SignedIDkgDealing`s.
Each signature in the batch is 64 bytes, therefore the largest share of
data is allocated to the dealings themselves.

The realization is that the `SignedIDkgDealing`s were already
broadcasted to all nodes as part of the IDKG protocol itself. They
should therefore already exist in the IDKG pool of each peer. In fact,
including or seeing a batch-signed dealing as part of a transcript is
proof that at least `2f+1 `nodes have seen, verified and supported the
dealing already.

Consequently, there is no need to broadcast the dealings a second time
as part of the block proposal. This allows us to effectively re-use the
idea of "hashes-in-blocks", and replace the `SignedIDkgDealing`s in the
transcript with hashes (IDs), before transmitting the block over the
network.

## Changes
When receiving such a "stripped" block proposal (with artifact IDs
instead of full dealings), a node might not have all dealings
corresponding to all of the included IDs locally. In that case, it needs
to request the missing dealings from other peers.

This PR adds a new endpoint for requesting `SignedIDkgDealing`s by the
artifact ID and block ID. When receiving such a request, the node will
first attempt to find the requested dealing in its own IDKG pool. If it
doesn't exist, then it will attempt to find it in the block with the
specified block proposal ID.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>