chore: Remove async from partition tools
Async functions are useful because we run multiple async tasks on the same thread. However, they also introduce additional complexity. The code that uses the utils in partition_tools doesn't have to handle many parallel tasks. In fact, we usually run a single task at a time so we can't profit from the benefits of async.