feat(ckdoge): add `estimate_withdrawal_fee` query endpoint (#7801)
Add a query endpoint `estimate_withdrawal_fee` to estimate the cost of a
ckDOGE withdrawal, similar to the one in ckBTC.

The main changes are as follows:

1. Improve the fee charged in case some pending withdrawals cannot be
processed to account for the exact number of impacted withdrawal
requests instead of using the upper bound on the maximum number of
withdrawal that can be processed in one batch
([DEFI-2028)](https://dfinity.atlassian.net/browse/DEFI-2028).
2. Augment the `FeeEstimator` trait to account for the minimum fee rate
increase for transaction resubmission and computing the fee for
reimbursing a certain number of pending withdrawal requests.