feat: enabling the swaps for all node operators on non-system subnets (#7818)
This PR enables the node swapping feature for all node operators on
non-system subnets.

The feature was initially enabled just for Dfinity node operators for
testing (also only on non-system subnets) and this PR just allows node
operators to use it.

This is the second phase of the rollout of the swapping feature.