perf(ckbtc): canbench `build_unsigned_transaction_from_inputs` (#7782)
Benchmark the method to build unsigned transactions for various
withdrawal amounts by re-using the `canbench` infrastructure introduced
in #4240 to rebuild the minter's state from mainnet events.

As expected, the larger the amount, the more instructions are required
since the more UTXOs are needed to cover that withdrawal amount.
However, if the withdrawal amount remains in the range 50k sats (minimum
withdrawal amount) to 0.1 BTC, the number of instructions is similar;
while, there is a significant gap starting with withdrawal amount of 1
BTC:

1. 50k sats to 0.1 BTC: 15M instructions
2. 1 BTC: 400M instructions
3. 10 BTC: 5B instructions