fix: Fix checkpointing after snapshot upload
This commit fixes a problem with persisting uploaded canister snapshots.
Previously it was possible for uploaded data to be lost on
checkpointing, resulting in zeroed out canister data. The problem was
most likely to be observed if the snapshot is large and needed to be
uploaded across several checkpoint intervals, but even for small
canisters it could occur.

The checkpointing logic generally assumes that snapshots are immutable,
and no changes need to be persisted, unless a snapshot is marked as
dirty by the management canister by calling
`record_snapshot_data_upload`. This did however not happen, resulting in
a situation where data would get lost, unless the data upload happened
in the same checkpoint interval as the metadata upload.