chore: disable IPv4 for IcGatewayVm in tests (#7854)
There are issues with our DHCP server in our zh1 DC which is causing
deploying IcGatewayVms to fail since they request an IPv4 address which
times out.

The tests that use IcGatewayVms don't require them to have an IPv4
address so we disable IPv4 for IcGatewayVms in tests but leave it
enabled for IcGatewayVms in dynamic testnets.