feat(node-rewards-canister): Use async task for NRC hourly sync (#7865)
This pull request refactors the recurring timer task logic in
`timer_tasks.rs` to use an async trait for retrying sync with shorter
duration.

The main change is replacing the previous `RecurringSyncTask` trait
implementation with a new, non-send async trait.
Unfortunately is not possible to directly use RecurringAsyncTask because
MetricsManager and StableCanisterRegistryClient are not Send.

Making them Send requires quite extensive refactor that is left TODO.