chore: migrate rules_rust to MODULE.bazel (#7872)
This migrates our `rules_rust` setup from the legacy `WORKSPACE.bazel`
to `MODULE.bazel`. In particular:

* The toolchain and crate setup is moved to `bazel/rust.MODULE.bazel`
which is included from `MODULE.bazel`
* Two patches are introduced to bridge the gap in features between the
legacy WORKSPACE and the new MODULE rules_rust support
* Some labels had to be rewritten in canonical form (`@lmdb` ->
`@@_main~_repo_rules~lmdb`)
* The pinning script was simplified and support for pinning a single
crate was removed — didn't seem to be used but can be reverted if
necessary.
* Some canbench measurements were updated to reflect performance
improvements (as of right now unclear why)

Notes:
* the WORKSPACE was left for now as some Bazel targets use it as input
to mark the root of the repo, will be removed subsequently
* The fuzzing flags currently have to be copied because `MODULE.bazel`
does not allow `load`ing other macros. This will be addressed next.