test(orchestrator): compare latest computed root hash pre-upgrade with CUP state hash post-upgrade (#7525)
In cases where the orchestrator breaks after an upgrade, provisioning
readonly SSH keys would not be possible to recover the subnet. In that
case, there is no easy way to know the latest state hash to be included
in the recovery CUP except from hoping that the recovery operator's node
is up to date. Though, the state manager regularly logs the latest
computed state hash and a recovery operator could look at logs of all
nodes and use this information to build a recovery CUP.

This PR makes sure that this log is always observed **from the log
endpoint** before an upgrade, and compares it with the state hash that
the node reboots with. That way, we test that we can reliably observe
this log before a node reboots and we make sure that the test would
break if that log was ever removed in the future.

Furthermore, it always checks that the orchestrator gracefully exits.
This was not the case because back when the test was added, the relevant
log was not merged into mainnet yet, so we could only test it when
downgrading HEAD to mainnet and not the opposite.