test(ICRC_Index): DEFI-2527: index-ng u256 token testing (#7788)
Not all `index-ng` tests were also being run with `u256` tokens - in
particular, the integration tests under `tests/`, with the bazel target
`name = "index_ng_test" + conf["test_suffix"],`, were only using the
`u256` feature for one config, and not e.g., for the
`"get_blocks_disabled"` and `"icrc3_disabled"` features. This PR
proposes to add the `u256-tokens` feature flag where it was missing, and
also proposes to add a test that explicitly verifies that `Tokens::MAX`
returns the expected value, and verifies that the ledger and index
handle large transfer and approve amounts correctly.