refactor: drop AddCanisterChangeToHistory enum (#7860)
This PR drops the `AddCanisterChangeToHistory` enum by inlining the
canister history update to `CanisterManager::uninstall_code`.

The motivation for this change is to decouple `RoundLimits` update due
to uninstalling code and due to updating canister history: those steps
will be performed separately (in a follow-up PR) when taking a canister
snapshot and uninstalling its code atomically (and thus setting
`RoundLimits` to `None` would conveniently skip the `RoundLimits` update
in the helper function `uninstall_canister`).