chore: new fields for uninstalling code while taking canister snapshot (#7855)
This PR adds two extra fields to the input type of the
`take_canister_snapshot` method of the management canister:
- `uninstall_code`: allows to atomically uninstall the canister after
taking its snapshot; in particular, the canister memory usage is updated
atomically and thus it does not grow significantly (ignoring some
potential constant overhead for certified variables which are not
accounted for by canister memory usage, but are accounted for in
canister snapshot memory usage),
- `sender_canister_version`: needed to record the sender canister
version in the code uninstallation entry in canister history.

Warning! The actual functionality will only be implemented in follow-up
PRs.