feat: add DMZ as a HostFeature option (#7895)
For the NNS-recovery drill external Node Providers will need to connect
their IC nodes to a Farm-deployed testnet. In order for their hosts to
reach the testnet the testnet needs to be deployed into a DMZ to which
the public Internet has access.

Farm has been extended with some hosts in zh1 configured into a DMZ.
These hosts have `mandatoryFeatures = ["dmz"]` such that they only host
VMs that require the `dmz` feature. This commit adds the
`HostFeature::DMZ` such that testnets can require it.