chore(migration-canister): return only a single migration status (#7890)
This PR makes the `migration_status` endpoint of the migration canister
return only a single or no migration status for a given pair of canister
IDs.

This PR also fixes the duration of waiting for a completed migration in
the canister migration system test to 7 minutes (since the migration
canister now waits for at least 6 minutes before completing canister
migration to prevent replay attacks - see
[PR](https://github.com/dfinity/ic/pull/7787)).

To prevent such regressions in the future, this PR adds the system test
to `PULL_REQUEST_BAZEL_TARGETS` and tags the test as `long_test` so that
it can run on PRs using `PULL_REQUEST_BAZEL_TARGETS`.