perf(crypto): CRP-2926 adapt VetKdArgs to only hold references (#7356)
Adapts the `VetKdArgs` to only hold references instead of owning the
data.

This is done because (1) it is conceptually cleaner and (2) it
eliminates the clonings required in the Consensus layer (for caller,
context, ni_dkg_id, input, transport_public_key) to call the
vetKD-related methods in the Crypto Component, i.e.,
VetKdProtocol::create_encrypted_key_share,
VetKdProtocol::verify_encrypted_key_share,
VetKdProtocol::combine_encrypted_key_shares,
VetKdProtocol::verify_encrypted_key.

While the savings are minor (because there are no large structs
involved), this mainly unifies the approach also taken for
ThresholdEcdsaSigInputs (see https://github.com/dfinity/ic/pull/7224)
and ThresholdSchnorrSigInputs (see
https://github.com/dfinity/ic/pull/7285).

Note that flattening the derivation path from the
`SignWithThresholdContext` doesn't work any more because it would create
an intermediate value that we cannot reference because it is dropped at
the end of the function. Instead we use the `first()` entry in the
vector. Using `.first()` is correct, however, because for vetKD, the
execution layer creates a vector with just a single entry (see
[this](https://github.com/dfinity/ic/blob/fcc2d0f2c5b09e1c06dfdd99d647d2967152a9b7/rs/execution_environment/src/execution_environment.rs#L3346)).
To get around this issue in the first place (i.e., not get passed a
`Vec<Vec<u8>>` from the `SignWithThresholdContext`), we would have to
move the `SignWithThresholdContext::derivation_path` into the
`SignWithThresholdContext::args: ThresholdArguments` (where
`ThresholdArguments` is an enum over `Ecdsa`, `Schnorr`, `VetKd`).
Because the `SignWithThresholdContext` is part of the replicated state
in protobuf form, we’d have to do a migration first, which is involved
and time-consuming, so we chose the approach with using `first()`, at
least for now. The state machine tests were adapted accordingly to also
just use the first element.

Various consensus test utilities were adapted to use `vec![vec![]]`
instead of `vec![]` as derivation path in the `SignWithThresholdContext`
because this context is used for both tECDSA/tSchnorr and vetKD, and the
vetKD code requires now that there is exactly one entry in the
derivation path.

The `VetKdArgs` no longer implement `Deserialize, Serialize` because
`serde::Deserialize` cannot be implemented on such reference structs.

In the tests, an owned variant (`VetKdArgsOwned`) was created for
improved usability in tests so that testing functions can create and
return vetKD args.